requirements.txt
Django==2.2
django-cors-middleware==1.3.1
django-crispy-forms==1.7.2
django-filter==2.0.0
django-oauth-toolkit==1.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.1
Markdown==2.6.11
mysqlclient==1.3.13
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.5
whitenoise>=3.3.1
gunicorn==19.9.0
django-heroku==0.3.1